Movement of methods for single responsibility principle
Looking at the provided code, it seems like there are a few methods that could be moved to different classes to better align with the Single Responsibility Principle (SRP). The SRP states that a class should have only one reason to change, meaning it should only have one job or responsibility.  Here's a plan for moving some methods:  1. The  SignUpTopic  class has methods that deal with the  SignedUpTeam  model ( sign_team_up ,  assign_topic_to_team ,  save_waitlist_entry ). These methods should be moved to the  SignedUpTeam  model, as they are more about the state and behavior of a signed up team than about the topic itself.  2. Similarly, the  SignUpSheet  class has a method  signup_team  that seems to deal more with the state of a  SignedUpTeam  than the sign up sheet itself. This method should also be moved to the  SignedUpTeam  model.  3. The  ReviewResponseMap  class has a method  prepare_final_review _versions  that deals with the state of a  ResponseMap . This method should be moved to the  ResponseMap  model.  4. The  SignUpTopic  class has methods  longest_waiting_team  and  reassign_topic  that are more about the management of teams than the topic itself. These methods should be moved to the  Team  model.  5. The  SignUpTopic  class has a method  users_on_waiting_lis t  that seems to deal more with the state of a  SignedUpTeam  than the topic itself. This method should also be moved to the  SignedUpTeam  model.  Please confirm if this plan seems reasonable and I will proceed to write the code changes.

To improve clarity and adhere to the Single Responsibility Principle, you could move methods from  SignUpTopic  and  SignUpSheet  to the  SignedUpTeam  class, and also shift some methods from  ReviewResponseMap  to  ResponseMap . Does this plan sound good to you?